refactor: use dblink to CREATE DATABASE in sql script
this is a separate attempt to scaffold the tournament database using only the tournament.sql script executable. Unfortunately, PostgreSQL does not allow for 'CREATE DATABASE IF NOT EXISTS', so a workaround involving dblink has been implimented. The remaining issue with this method is that CREATE EXTENSION is not allowed for users who are not superusers (such as postgres). This is especially difficult for vm environments where the coder cannot have access to the os user that is linked to the postgres superuser, and is effectively locked out from the superuser role.

todo: include code to root and/or gain superuser authorization????